Save access token and refresh token in database after successful network calls to fetch an access token for the first time or refresh an access token using the refresh token

+++ app/src/main/java/com/example/guava/oauth/access/AccessTokenEntity.kt
@@ -14,5 +14,5 @@ data class AccessTokenEntity(
     @ColumnInfo(name = "scope")
     val scope: Boolean,
     @ColumnInfo(name = "expires_in")
-    val expiresInSecs: Int
+    val expiresInSecs: Long
 )
